large rehaul, removing more stuff, merging into main base issue reporter service now
