+++ src/vs/workbench/contrib/issue/browser/issueReporterService.ts
@@ -2,15 +2,10 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import { $, isHTMLInputElement, isHTMLTextAreaElement, reset, windowOpenNoOpener } from 'vs/base/browser/dom';
-import { Codicon } from 'vs/base/common/codicons';
-import { groupBy } from 'vs/base/common/collections';
-import { isMacintosh } from 'vs/base/common/platform';
 import { IProductConfiguration } from 'vs/base/common/product';
-import { ThemeIcon } from 'vs/base/common/themables';
 import { localize } from 'vs/nls';
-import { BaseIssueReporterService } from 'vs/workbench/contrib/issue/browser/issue';
-import { IIssueFormService, IssueReporterData, IssueReporterExtensionData } from 'vs/workbench/contrib/issue/common/issue';
+import { BaseIssueReporterService } from 'vs/workbench/contrib/issue/browser/baseIssueReporterService';
+import { IIssueFormService, IssueReporterData } from 'vs/workbench/contrib/issue/common/issue';
 
 // GitHub has let us know that we could up our limit here to 8k. We chose 7500 to play it safe.
 // ref https://github.com/microsoft/vscode/issues/159191
@@ -40,37 +35,11 @@ export class IssueWebReporter extends BaseIssueReporterService {
 		}
 
 		this.setEventHandlers();
-		this.handleExtensionData(data.enabledExtensions);
-	}
-
-	private handleExtensionData(extensions: IssueReporterExtensionData[]) {
-		const installedExtensions = extensions.filter(x => !x.isBuiltin);
-		const { nonThemes, themes } = groupBy(installedExtensions, ext => {
-			return ext.isTheme ? 'themes' : 'nonThemes';
-		});
-
-		const numberOfThemeExtesions = themes && themes.length;
-		this.issueReporterModel.update({ numberOfThemeExtesions, enabledNonThemeExtesions: nonThemes, allExtensions: installedExtensions });
-		this.updateExtensionTable(nonThemes, numberOfThemeExtesions);
-		if (this.disableExtensions || installedExtensions.length === 0) {
-			(<HTMLButtonElement>this.getElementById('disableExtensions')).disabled = true;
-		}
-
-		this.updateExtensionSelector(installedExtensions);
-	}
-
-	private async sendReporterMenu(extension: IssueReporterExtensionData): Promise<IssueReporterData | undefined> {
-		try {
-			const data = await this.issueFormService.sendReporterMenu(extension.id);
-			return data;
-		} catch (e) {
-			console.error(e);
-			return undefined;
-		}
 	}
 
 	public override setEventHandlers(): void {
 		super.setEventHandlers();
+
 		this.addEventListener('issue-type', 'change', (event: Event) => {
 			const issueType = parseInt((<HTMLInputElement>event.target).value);
 			this.issueReporterModel.update({ issueType: issueType });
@@ -84,160 +53,6 @@ export class IssueWebReporter extends BaseIssueReporterService {
 			this.updatePreviewButtonState();
 			this.setSourceOptions();
 			this.render();
-		});
-		this.previewButton.onDidClick(async () => {
-			this.delayedSubmit.trigger(async () => {
-				this.createIssue();
-			});
-		});
-
-		this.addEventListener('disableExtensions', 'click', () => {
-			this.issueFormService.reloadWithExtensionsDisabled();
-		});
-
-		this.addEventListener('extensionBugsLink', 'click', (e: Event) => {
-			const url = (<HTMLElement>e.target).innerText;
-			windowOpenNoOpener(url);
-		});
-
-		this.addEventListener('disableExtensions', 'keydown', (e: Event) => {
-			e.stopPropagation();
-			if ((e as KeyboardEvent).key === 'Enter' || (e as KeyboardEvent).key === ' ') {
-				this.issueFormService.reloadWithExtensionsDisabled();
-			}
-		});
-
-		this.window.document.onkeydown = async (e: KeyboardEvent) => {
-			const cmdOrCtrlKey = isMacintosh ? e.metaKey : e.ctrlKey;
-			// Cmd/Ctrl+Enter previews issue and closes window
-			if (cmdOrCtrlKey && e.key === 'Enter') {
-				this.delayedSubmit.trigger(async () => {
-					if (await this.createIssue()) {
-						this.close();
-					}
-				});
-			}
-
-			// Cmd/Ctrl + w closes issue window
-			if (cmdOrCtrlKey && e.key === 'w') {
-				e.stopPropagation();
-				e.preventDefault();
-
-				const issueTitle = (<HTMLInputElement>this.getElementById('issue-title'))!.value;
-				const { issueDescription } = this.issueReporterModel.getData();
-				if (!this.hasBeenSubmitted && (issueTitle || issueDescription)) {
-					// fire and forget
-					this.issueFormService.showConfirmCloseDialog();
-				} else {
-					this.close();
-				}
-			}
-
-			// With latest electron upgrade, cmd+a is no longer propagating correctly for inputs in this window on mac
-			// Manually perform the selection
-			if (isMacintosh) {
-				if (cmdOrCtrlKey && e.key === 'a' && e.target) {
-					if (isHTMLInputElement(e.target) || isHTMLTextAreaElement(e.target)) {
-						(<HTMLInputElement>e.target).select();
-					}
-				}
-			}
-		};
-	}
-
-	private updateExtensionSelector(extensions: IssueReporterExtensionData[]): void {
-		interface IOption {
-			name: string;
-			id: string;
-		}
-
-		const extensionOptions: IOption[] = extensions.map(extension => {
-			return {
-				name: extension.displayName || extension.name || '',
-				id: extension.id
-			};
-		});
-
-		// Sort extensions by name
-		extensionOptions.sort((a, b) => {
-			const aName = a.name.toLowerCase();
-			const bName = b.name.toLowerCase();
-			if (aName > bName) {
-				return 1;
-			}
-
-			if (aName < bName) {
-				return -1;
-			}
-
-			return 0;
-		});
-
-		const makeOption = (extension: IOption, selectedExtension?: IssueReporterExtensionData): HTMLOptionElement => {
-			const selected = selectedExtension && extension.id === selectedExtension.id;
-			return $<HTMLOptionElement>('option', {
-				'value': extension.id,
-				'selected': selected || ''
-			}, extension.name);
-		};
-
-		const extensionsSelector = this.getElementById<HTMLSelectElement>('extension-selector');
-		if (extensionsSelector) {
-			const { selectedExtension } = this.issueReporterModel.getData();
-			reset(extensionsSelector, this.makeOption('', localize('selectExtension', "Select extension"), true), ...extensionOptions.map(extension => makeOption(extension, selectedExtension)));
-
-			if (!selectedExtension) {
-				extensionsSelector.selectedIndex = 0;
-			}
-
-			this.addEventListener('extension-selector', 'change', async (e: Event) => {
-				this.clearExtensionData();
-				const selectedExtensionId = (<HTMLInputElement>e.target).value;
-				this.selectedExtension = selectedExtensionId;
-				const extensions = this.issueReporterModel.getData().allExtensions;
-				const matches = extensions.filter(extension => extension.id === selectedExtensionId);
-				if (matches.length) {
-					this.issueReporterModel.update({ selectedExtension: matches[0] });
-					const selectedExtension = this.issueReporterModel.getData().selectedExtension;
-					if (selectedExtension) {
-						const iconElement = document.createElement('span');
-						iconElement.classList.add(...ThemeIcon.asClassNameArray(Codicon.loading), 'codicon-modifier-spin');
-						this.setLoading(iconElement);
-						const openReporterData = await this.sendReporterMenu(selectedExtension);
-						if (openReporterData) {
-							if (this.selectedExtension === selectedExtensionId) {
-								this.removeLoading(iconElement, true);
-								this.data = openReporterData;
-							} else if (this.selectedExtension !== selectedExtensionId) {
-							}
-						}
-						else {
-							if (!this.loadingExtensionData) {
-								iconElement.classList.remove(...ThemeIcon.asClassNameArray(Codicon.loading), 'codicon-modifier-spin');
-							}
-							this.removeLoading(iconElement);
-							this.clearExtensionData();
-							selectedExtension.data = undefined;
-							selectedExtension.uri = undefined;
-						}
-						if (this.selectedExtension === selectedExtensionId) {
-							// repopulates the fields with the new data given the selected extension.
-							this.updateExtensionStatus(matches[0]);
-							this.openReporter = false;
-						}
-					} else {
-						this.issueReporterModel.update({ selectedExtension: undefined });
-						this.clearSearchResults();
-						this.clearExtensionData();
-						this.validateSelectedExtension();
-						this.updateExtensionStatus(matches[0]);
-					}
-				}
-			});
-		}
-
-		this.addEventListener('problem-source', 'change', (_) => {
-			this.validateSelectedExtension();
 		});
 	}
 }
